Restructured reflection. applied pr feedback
