+++ src/lib.rs
@@ -47,7 +47,12 @@ pub mod wrapper;
 
 pub mod intellisense;
 
+#[cfg(windows)]
+pub mod dxil;
+#[cfg(windows)]
+pub use crate::dxil::validate_dxil;
+
 pub use crate::ffi::*;
 pub use crate::utils::compile_hlsl;
-pub use crate::utils::validate_dxil;
 pub use crate::wrapper::*;
+
